refactor: extract Levenshtein helper and drop redundant pre-stat in edit
Tools.ts's quoted-key suggester and EditMatcher's strategy similarity
score both had near-identical Levenshtein implementations. Promote one
copy to shared/Levenshtein.ts and import from both.

edit's earlier change pre-statted via FsErrors.statOrThrow purely to get
the friendly ENOENT message, then realpath'd anyway. realpath throws
ENOENT itself — wrap that catch instead, saves a syscall and removes a
small TOCTOU window where the file could disappear between stat and
realpath.
