+++ src/shared/Tools.ts
@@ -4,6 +4,7 @@ import type {
 } from "@earendil-works/pi-coding-agent";
 import { validateToolArguments } from "@earendil-works/pi-ai";
 import type { Static, TSchema } from "typebox";
+import { Levenshtein } from "./Levenshtein";
 
 type Issue = { readonly path: string; readonly message: string };
 
@@ -624,7 +625,7 @@ function closestKey(
     if (candidate.toLowerCase() === lowered) {
       return candidate;
     }
-    const d = levenshtein(lowered, candidate.toLowerCase());
+    const d = Levenshtein.distance(lowered, candidate.toLowerCase());
     if (d <= Math.max(2, Math.floor(candidate.length / 3))) {
       if (!best || d < best.distance) {
         best = { key: candidate, distance: d };
@@ -632,26 +633,4 @@ function closestKey(
     }
   }
   return best?.key;
-}
-
-function levenshtein(a: string, b: string): number {
-  if (a === b) {
-    return 0;
-  }
-  if (!a.length) {
-    return b.length;
-  }
-  if (!b.length) {
-    return a.length;
-  }
-  let prev = Array.from({ length: b.length + 1 }, (_, i) => i);
-  for (let i = 1; i <= a.length; i++) {
-    const curr = [i];
-    for (let j = 1; j <= b.length; j++) {
-      const cost = a[i - 1] === b[j - 1] ? 0 : 1;
-      curr.push(Math.min(curr[j - 1]! + 1, prev[j]! + 1, prev[j - 1]! + cost));
-    }
-    prev = curr;
-  }
-  return prev[b.length]!;
 }
